major refactoring of package structure
