+++ NeuralSolvers/__init__.py
@@ -1,41 +1,14 @@
-from . import models
-from .InitalCondition import InitialCondition
-from .BoundaryCondition import PeriodicBC
-from .BoundaryCondition import DirichletBC
-from .BoundaryCondition import RobinBC
-from .BoundaryCondition import TimeDerivativeBC
-from .BoundaryCondition import NeumannBC
-from .PDELoss import PDELoss
-from .HPMLoss import HPMLoss
-from .Logger_Interface import LoggerInterface
-from .WandB_Logger import WandbLogger
-from .TensorBoard_Logger import TensorBoardLogger
-from .PINN import PINN
-from .Random_Sampler import RandomSampler
-from .LHS_Sampler import LHSSampler
-from .Adaptive_Sampler import AdaptiveSampler
+
 from .ND_Cube import NDCube
+from .pinn.PINN import PINN
 
 import NeuralSolvers.models
 import NeuralSolvers.callbacks
-
+import NeuralSolvers.pinn
+import NeuralSolvers.samplers
+import NeuralSolvers.loggers
 
 __all__ = [
-    'InitialCondition',
-    'PeriodicBC',
-    'DirichletBC',
-    'RobinBC',
-    'NeumannBC',
-    'TimeDerivativeBC',
-    'PDELoss',
-    'RandomSampler',
-    'LHSSampler',
-    'AdaptiveSampler',
     'NDCube'
-    'HPMLoss',
-    'PINN',
-    'models',
-    'LoggerInterface',
-    'WandbLogger',
-    'TensorBoardLogger',
-    'callbacks']
+    'PINN'
+    ]
