--- NeuralSolvers/callbacks/__init__.py
@@ -1,6 +1,0 @@
-from .Callback import Callback
-from .Callback import CallbackList
-
-___all__ = ["Callback",
-            "CallbackList"
-            ]
