+++ NeuralSolvers/pinn/__init__.py
@@ -1,0 +1,11 @@
+from .PINN import PINN
+from .HPMLoss import HPMLoss
+from .PDELoss import PDELoss
+
+import NeuralSolvers.pinn.datasets
+
+__all__ = [
+    'PDELoss',
+    'HPMLoss',
+    'PINN',
+]
