+++ NeuralSolvers/pinn/datasets/InitalCondition.py
@@ -1,4 +1,4 @@
-from .LossTerm import LossTerm
+from NeuralSolvers.LossTerm import LossTerm
 from torch import Tensor
 from torch.nn import Module
