+++ NeuralSolvers/samplers/LHS_Sampler.py
@@ -1,16 +1,18 @@
 import torch
 import numpy as np
-from .Sampler import Sampler
+from pyDOE import lhs
+from NeuralSolvers.samplers.Sampler import Sampler
 
 
-class RandomSampler(Sampler):
-    def __init__(self):
+class LHSSampler(Sampler):
+    def __init__(self, device = 'cpu'):
         """
-        Constructor of the RandomSampler (pseudo random sampler) class       
+        Constructor of the LHSSampler class
         """
-        super(RandomSampler, self).__init__()        
+        super(LHSSampler, self).__init__()
+        self.device = device
 
-    def sample(self, lb, ub, n):
+    def sample(self, lb, ub, n):        
         """Generate 'n' number of sample points in [lb,ub]
         
         Args:
@@ -26,5 +28,8 @@ class RandomSampler(Sampler):
         ub =  ub.reshape(1,-1)
         
         dimension = lb.shape[1]
-        xf = np.random.uniform(lb,ub,size=(n, dimension))
-        return torch.tensor(xf).float()
+        xf = lb + (ub - lb) * lhs(dimension, n)
+        xf_torch = torch.tensor(xf).float().to(self.device)
+
+
+        return xf_torch
