+++ examples/Burgers_Equation_1d/burgers_equation.png
