+++ examples/Heat_Equation_1d/Analytical_Solution.png
