+++ examples/Heat_Equation_1d/Analytical_different_t.png
