+++ examples/Heat_Equation_1d/Heat_Equation_adaptive.py
@@ -6,28 +6,19 @@ from torch.utils.data import Dataset
 import matplotlib.pyplot as plt
 import NeuralSolvers as nsolv
 
+from Heat_Equation import heat1d
+
 # Constants
-DEVICE = 'cpu'
-NUM_EPOCHS = 1000
+DEVICE = 'cpu' if torch.cuda.is_available() else 'cpu'
+NUM_EPOCHS = 10000
 DOMAIN_LOWER_BOUND = np.array([0, 0.0])
 DOMAIN_UPPER_BOUND = np.array([1.0, 2.0])
 NUM_INITIAL_POINTS = 50
 NUM_BOUNDARY_POINTS = 50
 NUM_COLLOCATION_POINTS = 20000
-
-
-def heat1d(x, u):
-    grads = ones(u.shape, device=u.device)
-    grad_u = grad(u, x, create_graph=True, grad_outputs=grads)[0]
-    u_x, u_t = grad_u[:, 0], grad_u[:, 1]
-
-    grads = ones(u_x.shape, device=u.device)
-    u_xx = grad(u_x, x, create_graph=True, grad_outputs=grads)[0][:, 0]
-
-    u_x, u_t, u_xx = [tensor.reshape(-1, 1) for tensor in (u_x, u_t, u_xx)]
-
-    return u_t - u_xx
-
+NUM_SEED_POINTS = 10000
+HIDDEN_SIZE = 100
+NUM_HIDDEN = 4
 
 class InitialConditionDataset(Dataset):
     def __init__(self, n0):
@@ -39,9 +30,7 @@ class InitialConditionDataset(Dataset):
         u = (np.exp(-alpha * t)) * np.sin(np.pi * x / L)
 
         idx = np.random.choice(x.shape[0], n0, replace=False)
-        self.x = x[idx, :]
-        self.u = u[idx, :]
-        self.t = t[idx, :]
+        self.x, self.u, self.t = x[idx, :], u[idx, :], t[idx, :]
 
     def __len__(self):
         return 1
@@ -50,7 +39,6 @@ class InitialConditionDataset(Dataset):
         x = np.concatenate([self.x, self.t], axis=1)
         y = self.u
         return Tensor(x).float(), Tensor(y).float()
-
 
 class BoundaryConditionDataset(Dataset):
     def __init__(self, nb, is_lower):
@@ -68,10 +56,9 @@ class BoundaryConditionDataset(Dataset):
     def __getitem__(self, idx):
         return Tensor(self.x_b).float()
 
-
 def setup_pinn():
     ic_dataset = InitialConditionDataset(n0=NUM_INITIAL_POINTS)
-    initial_condition = nsolv.InitialCondition(ic_dataset, name='Initial condition')
+    initial_condition = nsolv.pinn.datasets.InitialCondition(ic_dataset, name='Initial condition')
 
     bc_dataset_lb = BoundaryConditionDataset(nb=NUM_BOUNDARY_POINTS, is_lower=True)
     bc_dataset_ub = BoundaryConditionDataset(nb=NUM_BOUNDARY_POINTS, is_lower=False)
@@ -79,29 +66,33 @@ def setup_pinn():
     def dirichlet_func(x):
         return torch.zeros_like(x)[:, 0].reshape(-1, 1)
 
-    dirichlet_bc_lb = nsolv.DirichletBC(dirichlet_func, bc_dataset_lb, name='Lower dirichlet BC')
-    dirichlet_bc_ub = nsolv.DirichletBC(dirichlet_func, bc_dataset_ub, name='Upper dirichlet BC')
-
-    geometry = nsolv.NDCube(DOMAIN_LOWER_BOUND, DOMAIN_UPPER_BOUND, NUM_COLLOCATION_POINTS, NUM_COLLOCATION_POINTS,
-                            nsolv.LHSSampler(), device=DEVICE)
-
-    pde_loss = nsolv.PDELoss(geometry, heat1d, name='1D Heat', weight=1)
+    dirichlet_bc_lb = nsolv.pinn.datasets.DirichletBC(dirichlet_func, bc_dataset_lb, name='Lower dirichlet BC')
+    dirichlet_bc_ub = nsolv.pinn.datasets.DirichletBC(dirichlet_func, bc_dataset_ub, name='Upper dirichlet BC')
 
     model = nsolv.models.MLP(
         input_size=2, output_size=1, device=DEVICE,
-        hidden_size=100, num_hidden=4, lb=DOMAIN_LOWER_BOUND, ub=DOMAIN_UPPER_BOUND,
-        activation=torch.tanh
+        hidden_size=HIDDEN_SIZE, num_hidden=NUM_HIDDEN,
+        lb=DOMAIN_LOWER_BOUND, ub=DOMAIN_UPPER_BOUND
     )
 
-    return nsolv.PINN(model, 2, 1, pde_loss, initial_condition, [dirichlet_bc_lb, dirichlet_bc_ub], device=DEVICE)
+    sampler = nsolv.samplers.AdaptiveSampler(NUM_SEED_POINTS, model, heat1d)
+    geometry = nsolv.NDCube(DOMAIN_LOWER_BOUND, DOMAIN_UPPER_BOUND, NUM_COLLOCATION_POINTS, NUM_COLLOCATION_POINTS,
+                            sampler, device=DEVICE)
 
+    pde_loss = nsolv.pinn.PDELoss(geometry, heat1d, name='1D Heat')
 
-def train_pinn(pinn, num_epochs):
+    return nsolv.pinn.PINN(model, 2, 1, pde_loss, initial_condition, [dirichlet_bc_lb, dirichlet_bc_ub], device=DEVICE)
+
+def train_pinn(pinn):
+    '''logger = nsolv.WandbLogger("1D Heat equation pinn", {
+        "num_epochs": NUM_EPOCHS,
+        "hidden_size": HIDDEN_SIZE,
+        "num_hidden": NUM_HIDDEN
+    })
+    '''
     logger = None
-    #logger = nsolv.WandbLogger("1D Heat equation pinn", {"num_epochs": num_epochs})
-    pinn.fit(num_epochs, checkpoint_path='checkpoint.pt', restart=True, logger=logger,
-             lbfgs_finetuning=False, pretraining=True)
-
+    pinn.fit(NUM_EPOCHS, checkpoint_path='checkpoint.pt', restart=True,
+             logger=logger, lbfgs_finetuning=False, pretraining=True)
 
 def plot_solution(pinn):
     t = np.linspace(0, DOMAIN_UPPER_BOUND[1], 200).flatten()[:, None]
@@ -119,9 +110,8 @@ def plot_solution(pinn):
     plt.ylabel('x (cm)')
     plt.xlabel('t (seconds)')
     plt.colorbar().set_label('Temperature (°C)')
-    plt.title("PINN Solution: 1D Heat Equation")
+    plt.title("PINN Solution: 1D Heat Equation (Adaptive)")
     plt.show()
-
 
 def plot_analytical_solution():
     L, c = 1, 1
@@ -138,12 +128,13 @@ def plot_analytical_solution():
     plt.ylabel('x (cm)')
     plt.xlabel('t (seconds)')
     plt.colorbar().set_label('Temperature (°C)')
+
     plt.title("Analytical Solution: 1D Heat Equation")
     plt.show()
 
-
 if __name__ == "__main__":
     pinn = setup_pinn()
-    train_pinn(pinn, NUM_EPOCHS)
+    train_pinn(pinn)
+    pinn.load_model('best_model_pinn.pt')
     plot_solution(pinn)
     plot_analytical_solution()
