+++ examples/Heat_Equation_1d/Initial_condition.png
