+++ examples/Heat_Equation_1d/PDE_residual.png
