+++ examples/Heat_Equation_1d/PINN.png
