+++ examples/Heat_Equation_1d/PINN_different_t.png
