+++ examples/Heat_Equation_1d/PINNvsSol_100.png
