+++ examples/Heat_Equation_1d/PINNvsSol_199.png
