+++ examples/Heat_Equation_1d/PINNvsSol_50.png
