+++ examples/Schroedinger_1d/NLS.mat
